Update pic32prog.c
this is a HORRIBLE HACK, to prevent validating the baud rate (under linux) if the bitbang/'ascii ICSP' protocol is selected. this is necessary because the baud rate is repurposed by adapter-bitbang.c as a command to select several special functions:
-b1 to -b4 trigger upload of 'ascii ICSP' firmware to an attached Arduino NANO
-b8 and -b9 trigger a 'blind' erase of a uncooperative PIC32MX or PIC32MZ processor respectively

note that the 'ascii ICSP' protocol uses a fixed baud rate of 500000 that can not be changed, hence why the -b switch could be repurposed.

there is almost certainly a cleaner way to achieve the same result as this rather inelegant 1-line hack. please do feel free to reject this change and substitute a better alternative.

i've also found some other issues that are preventing 'ascii ICSP' from working under linux, where responses from the NANO seem to have characters missing. unfortunately i can't compile pic32prog for a linux target, so will have to resort to thinking through the issues to find a solution.

on a positive note, i have validated that the linux 32-bit compile works with a PICkit 2 under 32-bit linux mint, programming a PIC32MX170F256B quite successfully.

cheers,
rob   :-)